Adding a find item function to settings
this.settings.findItem will have the default behaviour, but can be customised for the user if the element to get `marked` or `unmarked` is a sibling or not a parent for that matter. This is handy when some predefined templates and plugins for customised form controls are used so that you don't need to modify the HTML or plugin to put the validator to work. :)